requirements.txt
flask==3.0.3
python-dotenv==1.0.1
supabase==2.5.1
gunicorn==21.2.0
